+++ react/components/ShippingSimulator/Wrapper.js
@@ -114,6 +114,12 @@ const BaseShippingSimulatorWrapper = ({
 
   useEffect(() => {
     if (isMountedRef.current) {
+      isMountedRef.current = false
+    }
+  }, [seller])
+
+  useEffect(() => {
+    if (isMountedRef.current) {
       return
     }
 
@@ -160,7 +166,6 @@ const ShippingSimulatorWithOrderForm = ({
 
   return (
     <BaseShippingSimulatorWrapper
-      key={seller}
       skuId={skuId}
       seller={seller}
       country={country}
@@ -218,7 +223,6 @@ const ShippingSimulatorWrapper = props => {
   ) {
     return (
       <BaseShippingSimulatorWrapper
-        key={sellerId}
         country={country}
         skuId={skuId}
         seller={sellerId}